requirements.txt
discord.py~=2.3.2
websockets~=12.0
python-dotenv~=1.0.1
